refactor: :recycle: Refactored avatar managment inside a AvatarCropper component
