+++ src/components/auth/Signup/Signup.scss
@@ -4,23 +4,4 @@
     & Form {
         @include form;
     }
-    & input[type="file"] {
-        display: none;
-    }
-    & .crop-container {
-        position: relative;
-        margin: auto;
-        width: 300px;
-        height: 300px;
-        & .avoid-badge {
-            position: absolute;
-            top: -10px;
-            right: -10px;
-            cursor: pointer;
-        }
-    }
-    & .slider-box {
-        width: 300px;
-        margin: auto;
-    }
 }
