+++ src/pages/Profile/Profile.tsx
@@ -1,19 +1,13 @@
 import { Formik, Form, Field, ErrorMessage } from 'formik';
 import { useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import Cropper from 'react-easy-crop';
-import { Point, Area } from 'react-easy-crop/types';
-import getCroppedImg from '../../utils/canvasUtils';
-import { Slider, Badge, Button, Breadcrumbs } from '@mui/material';
-import ChangeCircleIcon from '@mui/icons-material/ChangeCircle';
-import HighlightOffTwoToneIcon from '@mui/icons-material/HighlightOffTwoTone';
-import AddAPhotoTwoToneIcon from '@mui/icons-material/AddAPhotoTwoTone';
+
 import { FormValidationService } from '../../utils/formValidationService';
 import { UserService } from '../../services/userService';
 import { setUser } from '../../redux/authSlice';
+import AvatarCropper from '../../components/avatarCropper/AvatarCropper';
 
 import './Profile.scss';
-
 
 const formValidationService = new FormValidationService();
 const userService = new UserService();
@@ -23,47 +17,11 @@ export default function Profile() {
     const authStatus = useSelector((state:any) => state.auth);
     const [base64image, setBase64image] = useState<string>('');
     const [passwordEdit, setPasswordEdit] = useState(false);
-    const [crop, setCrop] = useState<Point>({ x:0, y: 0 });
-    const [ zoom, setZoom ] = useState<number>(1);
-    const [ cropperImage, setCropperImage] = useState<string>('');
-
-    const onCropComplete = async (croppedArea: Area, croppedAreaPixels: Area) => {
-        const crop = await getCroppedImg(cropperImage, croppedAreaPixels)
-        if(crop) {
-            setBase64image(crop);
-        }
-    };
-
-    const askImageSelection = () => {
-        const inputEl = document.querySelector('#image') as HTMLInputElement;
-        inputEl.click();
-    };
-    
-    const onFileSelected = (event:any) => {
-        const reader = new FileReader();
-        reader.readAsDataURL(event.target.files[0]);
-        reader.onload = function () {
-            if(reader.result) {
-                setCropperImage(reader.result.toString());
-                setBase64image(reader.result.toString());
-            }
-        };
-        reader.onerror = function (error) {
-            console.log('Error: ', error);
-        };
-    };
 
     const togglePasswordEdit = () => {
         setPasswordEdit(!passwordEdit);
     };
 
-    const avoidImageEdition = () => {
-        if(document.getElementById('image')){
-            const elt = document.getElementById('image') as HTMLInputElement;
-            elt.value="";
-        }
-        setCropperImage('');
-    };
     return (
         <div className="Profile">
             <h2>Profile works !</h2>
@@ -85,7 +43,7 @@ export default function Profile() {
                     userService.updateProfile(values, authStatus.user.id)
                     .then(response => {
                         dispatch(setUser(response.result));
-                        avoidImageEdition();
+                        setBase64image('');
                     })
                     .catch(error => {
                         console.log(error);
@@ -136,53 +94,10 @@ export default function Profile() {
                                     />
                                     <ErrorMessage name="oldPassword" />
                                 </div>
-
                             </div>
                         }
                         <div className="formgroup-heading">Avatar :</div>
-                        {cropperImage && 
-                            <div className="crop-container">
-                                <Cropper
-                                    image={cropperImage}
-                                    crop={crop}
-                                    cropShape="round"
-                                    showGrid={false}
-                                    zoom={zoom}
-                                    aspect={1}
-                                    onCropChange={setCrop}
-                                    onCropComplete={onCropComplete}
-                                    onZoomChange={setZoom}
-                                />
-                                <div className="avoid-badge" title="annuler" onClick={avoidImageEdition} >
-                                    <HighlightOffTwoToneIcon sx={{ color: '#800101' }} />
-                                </div>
-                            </div>
-                        }
-                        {cropperImage && <div className="slider-box"><Slider
-                            value={zoom}
-                            min={1}
-                            max={3}
-                            step={0.1}
-                            aria-labelledby="zoom"
-                            onChange={(e, zoom) => setZoom(Number(zoom))}
-                        /></div>}   
-                        <Field type="file" id="image" name="image" onChange={onFileSelected} />
-                        {authStatus.user.picture && 
-                            <div className="avatar-editor">
-                                <img className="actual-avatar" src={authStatus.user.picture} alt="your actual avatar" />
-                            <Breadcrumbs>
-                                <ChangeCircleIcon sx={{ color: '#1616c4' }} onClick={askImageSelection} />
-                                <HighlightOffTwoToneIcon sx={{ color: '#800101' }}/>
-                            </Breadcrumbs>
-                            </div>
-                        }
-                        {!authStatus.user.picture && !cropperImage && <div>
-                                <span>aucun</span><br/>
-                                <Button variant="contained" startIcon={<AddAPhotoTwoToneIcon />} onClick={askImageSelection}>
-                                    Ajouter
-                                </Button>
-                            </div>
-                        }
+                        <AvatarCropper setImage={setBase64image} userImage={authStatus.user.picture} />
                         <br />
                         <br />
                         <br />
@@ -190,7 +105,6 @@ export default function Profile() {
                     </Form>
                 )}
             </Formik>
-            
         </div>
     )
 }
